Bring back USNF's scripts-only mode for older Android, but better
- do scripts-only mode automatically on Android < 8.0 since it doesn't support Zygisk and generally doesn't need a PIF module to pass at all
- clean up all unused files in scripts-only mode, and modify script behaviors for it, including allowing GMS on DenyList, and even add it if missing since it's necessary on some scripts-only configurations
- allow manually entering scripts-only mode by deleting the zygisk directory of the currently installed module and reinstalling
- intentionally don't alter the Enforce DenyList state, since we don't know what other modules are installed that might depend on it one way or the other, so leave that up to the user

+++ module/post-fs-data.sh
@@ -1,13 +1,17 @@
 MODPATH="${0%/*}"
 . $MODPATH/common_func.sh
 
-# Remove Play Services from Magisk Denylist when set to enforcing
-if magisk --denylist status; then
-    magisk --denylist rm com.google.android.gms
+if [ -d "$MODPATH/zygisk" ]; then
+    # Remove Play Services from Magisk DenyList when set to Enforce in normal mode
+    if magisk --denylist status; then
+        magisk --denylist rm com.google.android.gms
+    fi
+    # Run common tasks for installation and boot-time
+    . $MODPATH/common_setup.sh
+else
+    # Add Play Services DroidGuard process to Magisk DenyList for better results in scripts-only mode
+    magisk --denylist add com.google.android.gms com.google.android.gms.unstable
 fi
-
-# Run common tasks for installation and boot-time
-. $MODPATH/common_setup.sh
 
 # Conditional early sensitive properties
 
@@ -26,7 +30,7 @@ resetprop_if_diff ro.boot.realmebootstate green
 # OnePlus
 resetprop_if_diff ro.is_ever_orange 0
 
-# Microsoft, RootBeer
+# Microsoft
 resetprop_if_diff ro.build.tags release-keys
 
 # Other
